+++ HealthTracker/Controller/HealthTrackerTabBarController.m
@@ -54,7 +54,7 @@
     {
         bmiCount = 0.0;
     }
-    tbi.badgeValue = [NSString stringWithFormat:@"%d",(NSInteger)bmiCount];//TODO: Get accurate BMI count
+    tbi.badgeValue = [NSString stringWithFormat:@"%ld",(long)bmiCount];//TODO: Get accurate BMI count
 }
 
 - (void)didReceiveMemoryWarning
